backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openpyxl==3.1.2
pandas==2.1.4
python-multipart==0.0.6
openai>=1.12.0
anthropic>=0.18.0
python-dotenv==1.0.0
email-validator==2.1.0
